+++ core/src/main/kotlin/net/corda/core/transactions/LedgerTransaction.kt
@@ -74,7 +74,13 @@ private constructor(
     private var componentGroups: List<ComponentGroup>? = null
     private var serializedInputs: List<SerializedStateAndRef>? = null
     private var serializedReferences: List<SerializedStateAndRef>? = null
-    private var isAttachmentTrusted: (Attachment) -> Boolean = { isAttachmentTrusted(it, null) }
+    private var isAttachmentTrusted: (Attachment) -> Boolean = {
+        when (it) {
+            is ContractAttachment -> isUploaderTrusted(it.uploader)
+            is AbstractAttachment -> isUploaderTrusted(it.uploader)
+            else -> false
+        }
+    }
 
     init {
         if (timeWindow != null) check(notary != null) { "Transactions with time-windows must be notarised" }
