+++ node/src/test/kotlin/net/corda/node/services/persistence/NodeAttachmentServiceTest.kt
@@ -763,356 +763,6 @@ class NodeAttachmentServiceTest {
     }
 
     @Test
-    fun `Jar uploaded by trusted uploader is trusted`() {
-        SelfCleaningDir().use { file ->
-            val (jar, _) = makeTestSignedContractJar(file.path, "foo.bar.DummyContract")
-            val unsignedJar = ContractJarTestUtils.makeTestContractJar(file.path, "com.example.MyContract")
-            val (attachment, _) = makeTestJar()
-
-            val signedId = jar.read { storage.privilegedImportAttachment(it, "app", "signed-contract.jar")}
-            val unsignedId = unsignedJar.read { storage.privilegedImportAttachment(it, "app", "unsigned-contract.jar") }
-            val attachmentId = attachment.read { storage.privilegedImportAttachment(it, "app", "attachment.jar")}
-
-            assertTrue(isAttachmentTrusted(storage.openAttachment(signedId)!!, storage), "Signed contract $signedId should be trusted but isn't")
-            assertTrue(isAttachmentTrusted(storage.openAttachment(unsignedId)!!, storage), "Unsigned contract $unsignedId should be trusted but isn't")
-            assertTrue(isAttachmentTrusted(storage.openAttachment(attachmentId)!!, storage), "Attachment $attachmentId should be trusted but isn't")
-        }
-    }
-
-    @Test
-    fun `jar trusted if signed by same key and has same contract as existing jar`() {
-        SelfCleaningDir().use { file ->
-            val alias = "testAlias"
-            val password = "testPassword"
-            val jarV1 = makeTestContractJar(file.path, "foo.bar.DummyContract")
-            file.path.generateKey(alias, password)
-            val key1 = file.path.signJar(jarV1.toAbsolutePath().toString(), alias, password)
-            val jarV2 = makeTestContractJar(file.path, "foo.bar.DummyContract", version = 2)
-            val key2 = file.path.signJar(jarV2.toAbsolutePath().toString(), alias, password)
-
-            val v1Id = jarV1.read { storage.privilegedImportAttachment(it, "app", "dummy-contract.jar") }
-            val v2Id = jarV2.read { storage.privilegedImportAttachment(it, "untrusted", "dummy-contract.jar") }
-
-            // Sanity check.
-            assertEquals(key1, key2, "Different public keys used to sign jars")
-            assertTrue(isAttachmentTrusted(storage.openAttachment(v1Id)!!, storage), "Initial contract $v1Id should be trusted")
-            assertTrue(isAttachmentTrusted(storage.openAttachment(v2Id)!!, storage), "Upgraded contract $v2Id should be trusted")
-        }
-    }
-
-    @Test
-    fun `jar trusted if same key but different contract`() {
-        SelfCleaningDir().use { file ->
-            val alias = "testAlias"
-            val password = "testPassword"
-            val jarV1 = makeTestContractJar(file.path, "foo.bar.DummyContract")
-            file.path.generateKey(alias, password)
-            val key1 = file.path.signJar(jarV1.toAbsolutePath().toString(), alias, password)
-            val jarV2 = makeTestContractJar(file.path, "foo.bar.DifferentContract", version = 2)
-            val key2 = file.path.signJar(jarV2.toAbsolutePath().toString(), alias, password)
-
-            val v1Id = jarV1.read { storage.privilegedImportAttachment(it, "app", "dummy-contract.jar") }
-            val v2Id = jarV2.read { storage.privilegedImportAttachment(it, "untrusted", "dummy-contract.jar") }
-
-            // Sanity check.
-            assertEquals(key1, key2, "Different public keys used to sign jars")
-            assertTrue(isAttachmentTrusted(storage.openAttachment(v1Id)!!, storage), "Initial contract $v1Id should be trusted")
-            assertTrue(isAttachmentTrusted(storage.openAttachment(v2Id)!!, storage), "Upgraded contract $v2Id should be trusted")
-        }
-    }
-
-    @Test
-    fun `jar trusted if the signing keys are a subset of an existing trusted jar's signers`() {
-        SelfCleaningDir().use { file ->
-            val alias = "testAlias"
-            val password = "testPassword"
-            val alias2 = "anotherTestAlias"
-            file.path.generateKey(alias, password)
-            file.path.generateKey(alias2, password)
-
-            val jarV1 = makeTestContractJar(file.path, "foo.bar.DummyContract")
-            file.path.signJar(jarV1.toAbsolutePath().toString(), alias, password)
-            file.path.signJar(jarV1.toAbsolutePath().toString(), alias2, password)
-
-            val jarV2 = makeTestContractJar(file.path, "foo.bar.DifferentContract", version = 2)
-            file.path.signJar(jarV2.toAbsolutePath().toString(), alias, password)
-
-            val v1Id = jarV1.read { storage.privilegedImportAttachment(it, "app", "dummy-contract.jar") }
-            val v2Id = jarV2.read { storage.privilegedImportAttachment(it, "untrusted", "dummy-contract.jar") }
-
-            assertTrue(isAttachmentTrusted(storage.openAttachment(v1Id)!!, storage), "Initial contract $v1Id should be trusted")
-            assertTrue(isAttachmentTrusted(storage.openAttachment(v2Id)!!, storage), "Upgraded contract $v2Id should be trusted")
-        }
-    }
-
-    @Test
-    fun `jar trusted if the signing keys are an intersection of an existing trusted jar's signers`() {
-        SelfCleaningDir().use { file ->
-            val alias = "testAlias"
-            val password = "testPassword"
-            val alias2 = "anotherTestAlias"
-            val alias3 = "yetAnotherTestAlias"
-            file.path.generateKey(alias, password)
-            file.path.generateKey(alias2, password)
-            file.path.generateKey(alias3, password)
-
-            val jarV1 = makeTestContractJar(file.path, "foo.bar.DummyContract")
-            file.path.signJar(jarV1.toAbsolutePath().toString(), alias, password)
-            file.path.signJar(jarV1.toAbsolutePath().toString(), alias2, password)
-
-            val jarV2 = makeTestContractJar(file.path, "foo.bar.DifferentContract", version = 2)
-            file.path.signJar(jarV2.toAbsolutePath().toString(), alias, password)
-            file.path.signJar(jarV2.toAbsolutePath().toString(), alias3, password)
-
-            val v1Id = jarV1.read { storage.privilegedImportAttachment(it, "app", "dummy-contract.jar") }
-            val v2Id = jarV2.read { storage.privilegedImportAttachment(it, "untrusted", "dummy-contract.jar") }
-
-            assertTrue(isAttachmentTrusted(storage.openAttachment(v1Id)!!, storage), "Initial contract $v1Id should be trusted")
-            assertTrue(isAttachmentTrusted(storage.openAttachment(v2Id)!!, storage), "Upgraded contract $v2Id should be trusted")
-        }
-    }
-
-    @Test
-    fun `jar trusted if the signing keys are a superset of an existing trusted jar's signers`() {
-        SelfCleaningDir().use { file ->
-            val alias = "testAlias"
-            val password = "testPassword"
-            val alias2 = "anotherTestAlias"
-            file.path.generateKey(alias, password)
-            file.path.generateKey(alias2, password)
-
-            val jarV1 = makeTestContractJar(file.path, "foo.bar.DummyContract")
-            file.path.signJar(jarV1.toAbsolutePath().toString(), alias, password)
-
-            val jarV2 = makeTestContractJar(file.path, "foo.bar.DifferentContract", version = 2)
-            file.path.signJar(jarV2.toAbsolutePath().toString(), alias, password)
-            file.path.signJar(jarV2.toAbsolutePath().toString(), alias2, password)
-
-            val v1Id = jarV1.read { storage.privilegedImportAttachment(it, "app", "dummy-contract.jar") }
-            val v2Id = jarV2.read { storage.privilegedImportAttachment(it, "untrusted", "dummy-contract.jar") }
-
-            assertTrue(isAttachmentTrusted(storage.openAttachment(v1Id)!!, storage), "Initial contract $v1Id should be trusted")
-            assertTrue(isAttachmentTrusted(storage.openAttachment(v2Id)!!, storage), "Upgraded contract $v2Id should be trusted")
-        }
-    }
-
-    @Test
-    fun `jar with inherited trust does not grant trust to other jars (no chain of trust)`() {
-        SelfCleaningDir().use { file ->
-            val aliasA = "Daredevil"
-            val aliasB = "The Punisher"
-            val aliasC = "Jessica Jones"
-            val password = "i am a netflix series"
-            file.path.generateKey(aliasA, password)
-            file.path.generateKey(aliasB, password)
-            file.path.generateKey(aliasC, password)
-
-            val jarSignedByA = makeTestContractJar(file.path, "foo.bar.DummyContract")
-            file.path.signJar(jarSignedByA.toAbsolutePath().toString(), aliasA, password)
-
-            val jarSignedByAB = makeTestContractJar(file.path, "foo.bar.DifferentContract", version = 2)
-            file.path.signJar(jarSignedByAB.toAbsolutePath().toString(), aliasB, password)
-            file.path.signJar(jarSignedByAB.toAbsolutePath().toString(), aliasA, password)
-
-            val jarSignedByBC = makeTestContractJar(file.path, "foo.bar.AnotherContract", version = 2)
-            file.path.signJar(jarSignedByBC.toAbsolutePath().toString(), aliasB, password)
-            file.path.signJar(jarSignedByBC.toAbsolutePath().toString(), aliasC, password)
-
-            val attachmentA = jarSignedByA.read { storage.privilegedImportAttachment(it, "app", "dummy-contract.jar") }
-            val attachmentB = jarSignedByAB.read { storage.privilegedImportAttachment(it, "untrusted", "dummy-contract.jar") }
-            val attachmentC = jarSignedByBC.read { storage.privilegedImportAttachment(it, "untrusted", "dummy-contract.jar") }
-
-            assertTrue(isAttachmentTrusted(storage.openAttachment(attachmentA)!!, storage), "Contract $attachmentA should be trusted")
-            assertTrue(isAttachmentTrusted(storage.openAttachment(attachmentB)!!, storage), "Contract $attachmentB should inherit trust")
-            assertFalse(isAttachmentTrusted(storage.openAttachment(attachmentC)!!, storage), "Contract $attachmentC should not be trusted (no chain of trust)")
-        }
-    }
-
-    @Test
-    fun `jar not trusted if different key but same contract`() {
-        SelfCleaningDir().use { file ->
-            val alias = "testAlias"
-            val password = "testPassword"
-            val jarV1 = makeTestContractJar(file.path, "foo.bar.DummyContract")
-            file.path.generateKey(alias, password)
-            val key1 = file.path.signJar(jarV1.toAbsolutePath().toString(), alias, password)
-            (file.path / "_shredder").delete()
-            (file.path / "_teststore").delete()
-            file.path.generateKey(alias, password)
-            val jarV2 = makeTestContractJar(file.path, "foo.bar.DummyContract", version = 2)
-            val key2 = file.path.signJar(jarV2.toAbsolutePath().toString(), alias, password)
-
-            val v1Id = jarV1.read { storage.privilegedImportAttachment(it, "app", "dummy-contract.jar") }
-            val v2Id = jarV2.read { storage.privilegedImportAttachment(it, "untrusted", "dummy-contract.jar") }
-
-            // Sanity check.
-            assertNotEquals(key1, key2, "Same public keys used to sign jars")
-            assertTrue(isAttachmentTrusted(storage.openAttachment(v1Id)!!, storage), "Initial contract $v1Id should be trusted")
-            assertFalse(isAttachmentTrusted(storage.openAttachment(v2Id)!!, storage), "Upgraded contract $v2Id should not be trusted")
-        }
-    }
-
-    @Test
-    fun `neither jar trusted if same contract and signer but not uploaded by a trusted uploader`() {
-        SelfCleaningDir().use { file ->
-            val alias = "testAlias"
-            val password = "testPassword"
-            val jarV1 = makeTestContractJar(file.path, "foo.bar.DummyContract")
-            file.path.generateKey(alias, password)
-            val key1 = file.path.signJar(jarV1.toAbsolutePath().toString(), alias, password)
-            val jarV2 = makeTestContractJar(file.path, "foo.bar.DummyContract", version = 2)
-            val key2 = file.path.signJar(jarV2.toAbsolutePath().toString(), alias, password)
-
-            val v1Id = jarV1.read { storage.privilegedImportAttachment(it, "untrusted", "dummy-contract.jar") }
-            val v2Id = jarV2.read { storage.privilegedImportAttachment(it, "untrusted", "dummy-contract.jar") }
-
-            // Sanity check.
-            assertEquals(key1, key2, "Different public keys used to sign jars")
-            assertFalse(isAttachmentTrusted(storage.openAttachment(v1Id)!!, storage), "Initial contract $v1Id should not be trusted")
-            assertFalse(isAttachmentTrusted(storage.openAttachment(v2Id)!!, storage), "Upgraded contract $v2Id should not be trusted")
-        }
-    }
-
-    @Test
-    fun `non contract jar trusted if trusted jar with same key present`() {
-        SelfCleaningDir().use { file ->
-            val alias = "testAlias"
-            val password = "testPassword"
-
-            // Directly use the ContractJarTestUtils version of makeTestJar to ensure jars are created in the right place, in order to sign
-            // them.
-            var counter = 0
-            val jarV1 = file.path / "$counter.jar"
-            ContractJarTestUtils.makeTestJar(jarV1.outputStream())
-            counter++
-            val jarV2 = file.path / "$counter.jar"
-            // Ensure that the first and second jars do not have the same hash
-            ContractJarTestUtils.makeTestJar(jarV2.outputStream(), entries = listOf(Pair("foo", "bar")))
-
-            file.path.generateKey(alias, password)
-            val key1 = file.path.signJar(jarV1.toAbsolutePath().toString(), alias, password)
-            val key2 = file.path.signJar(jarV2.toAbsolutePath().toString(), alias, password)
-
-            val v1Id = jarV1.read { storage.privilegedImportAttachment(it, "app", "dummy-attachment.jar") }
-            val v2Id = jarV2.read { storage.privilegedImportAttachment(it, "untrusted", "dummy-attachment-2.jar") }
-
-            // Sanity check.
-            assertEquals(key1, key2, "Different public keys used to sign jars")
-            assertTrue(isAttachmentTrusted(storage.openAttachment(v1Id)!!, storage), "Initial attachment $v1Id should be trusted")
-            assertTrue(isAttachmentTrusted(storage.openAttachment(v2Id)!!, storage), "Other attachment $v2Id should be trusted")
-        }
-    }
-
-    @Test
-    fun `all non contract jars not trusted if all are uploaded by non trusted uploaders`() {
-        SelfCleaningDir().use { file ->
-            val alias = "testAlias"
-            val password = "testPassword"
-
-            // Directly use the ContractJarTestUtils version of makeTestJar to ensure jars are created in the right place, in order to sign
-            // them.
-            var counter = 0
-            val jarV1 = file.path / "$counter.jar"
-            ContractJarTestUtils.makeTestJar(jarV1.outputStream())
-            counter++
-            val jarV2 = file.path / "$counter.jar"
-            // Ensure that the first and second jars do not have the same hash
-            ContractJarTestUtils.makeTestJar(jarV2.outputStream(), entries = listOf(Pair("foo", "bar")))
-
-            file.path.generateKey(alias, password)
-            val key1 = file.path.signJar(jarV1.toAbsolutePath().toString(), alias, password)
-            val key2 = file.path.signJar(jarV2.toAbsolutePath().toString(), alias, password)
-
-            val v1Id = jarV1.read { storage.privilegedImportAttachment(it, "untrusted", "dummy-attachment.jar") }
-            val v2Id = jarV2.read { storage.privilegedImportAttachment(it, "untrusted", "dummy-attachment-2.jar") }
-
-            // Sanity check.
-            assertEquals(key1, key2, "Different public keys used to sign jars")
-            assertFalse(isAttachmentTrusted(storage.openAttachment(v1Id)!!, storage), "Initial attachment $v1Id should not be trusted")
-            assertFalse(isAttachmentTrusted(storage.openAttachment(v2Id)!!, storage), "Other attachment $v2Id should not be trusted")
-        }
-    }
-
-    @Test
-    fun `non contract jars not trusted if unsigned`() {
-        SelfCleaningDir().use {
-            val (jarV1, _) = makeTestJar()
-            val (jarV2, _) = makeTestJar(entries = listOf(Pair("foo", "bar")))
-
-            val v1Id = jarV1.read { storage.privilegedImportAttachment(it, "app", "dummy-attachment.jar") }
-            val v2Id = jarV2.read { storage.privilegedImportAttachment(it, "untrusted", "dummy-attachment-2.jar") }
-
-            assertTrue(isAttachmentTrusted(storage.openAttachment(v1Id)!!, storage), "Initial attachment $v1Id should not be trusted")
-            assertFalse(isAttachmentTrusted(storage.openAttachment(v2Id)!!, storage), "Other attachment $v2Id should not be trusted")
-        }
-    }
-
-    @Test
-    fun `jar not trusted if signed by a blacklisted key and not uploaded by trusted uploader`() {
-        SelfCleaningDir().use { file ->
-
-            val aliasA = "Antman"
-            val aliasB = "The Wasp"
-            val password = "antman and the wasp"
-            file.path.generateKey(aliasA, password)
-            val keyB = file.path.generateKey(aliasB, password)
-
-            storage = NodeAttachmentService(
-                MetricRegistry(),
-                TestingNamedCacheFactory(),
-                database,
-                keysToBlacklist = listOf(keyB.hash.toString())
-            ).also {
-                database.transaction {
-                    it.start()
-                }
-            }
-            storage.servicesForResolution = services
-
-            val jarA = makeTestContractJar(file.path, "foo.bar.DummyContract")
-            file.path.signJar(jarA.toAbsolutePath().toString(), aliasA, password)
-            file.path.signJar(jarA.toAbsolutePath().toString(), aliasB, password)
-            val jarB = makeTestContractJar(file.path, "foo.bar.AnotherDummyContract")
-            file.path.signJar(jarB.toAbsolutePath().toString(), aliasA, password)
-            file.path.signJar(jarB.toAbsolutePath().toString(), aliasB, password)
-
-            val attachmentA = jarA.read { storage.privilegedImportAttachment(it, "app", "dummy-contract.jar") }
-            val attachmentB = jarB.read { storage.privilegedImportAttachment(it, "untrusted", "dummy-contract.jar") }
-
-            assertTrue(isAttachmentTrusted(storage.openAttachment(attachmentA)!!, storage), "Contract $attachmentA should be trusted")
-            assertFalse(isAttachmentTrusted(storage.openAttachment(attachmentB)!!, storage), "Contract $attachmentB should not be trusted")
-        }
-    }
-
-    @Test
-    fun `jar uploaded by trusted uploader is still trusted even if it is signed by a blacklisted key`() {
-        SelfCleaningDir().use { file ->
-
-            val aliasA = "Thanos"
-            val password = "what did it cost? everything"
-            val key = file.path.generateKey(aliasA, password)
-
-            storage = NodeAttachmentService(
-                MetricRegistry(),
-                TestingNamedCacheFactory(),
-                database,
-                keysToBlacklist = listOf(key.hash.toString())
-            ).also {
-                database.transaction {
-                    it.start()
-                }
-            }
-            storage.servicesForResolution = services
-
-            val jar = makeTestContractJar(file.path, "foo.bar.DummyContract")
-            file.path.signJar(jar.toAbsolutePath().toString(), aliasA, password)
-            val attachment = jar.read { storage.privilegedImportAttachment(it, "app", "dummy-contract.jar") }
-
-            assertTrue(isAttachmentTrusted(storage.openAttachment(attachment)!!, storage), "Contract $attachment should be trusted")
-        }
-    }
-
-    @Test
     fun `attachments can be queried by providing a intersection of signers using an EQUAL statement - EQUAL containing a single public key`() {
         SelfCleaningDir().use { file ->
             val aliasA = "Luke Skywalker"
@@ -1288,114 +938,6 @@ class NodeAttachmentServiceTest {
             )
             assertThat(result.map { (_, attachment) -> attachment.signerKeys }).containsOnly(
                 listOf(key), emptyList(), emptyList(), emptyList()
-            )
-        }
-    }
-
-    @Test
-    fun `resolveAttachmentTrustRoots returns all attachment trust roots`() {
-        SelfCleaningDir().use { file ->
-            val aliasA = "dan"
-            val aliasB = "james"
-            val password = "one day the attachment service will be refactored"
-            file.path.generateKey(aliasA, password)
-            file.path.generateKey(aliasB, password)
-
-            val jarSignedByA = makeTestContractJar(file.path, "foo.bar.AnotherContract")
-            file.path.signJar(jarSignedByA.toAbsolutePath().toString(), aliasA, password)
-
-            val jarSignedByAB = makeTestContractJar(file.path, "foo.bar.DummyContract")
-            file.path.signJar(jarSignedByAB.toAbsolutePath().toString(), aliasA, password)
-            file.path.signJar(jarSignedByAB.toAbsolutePath().toString(), aliasB, password)
-
-            val (zipC, hashB) = makeTestJar(listOf(Pair("file", "content")))
-            val (zipD, hashC) = makeTestJar(listOf(Pair("magic_file", "magic_content_puff")))
-
-            val attachmentA = jarSignedByA.read { storage.privilegedImportAttachment(it, "app", "A.jar") }
-            val attachmentB = jarSignedByAB.read { storage.privilegedImportAttachment(it, "untrusted", "B.jar") }
-            val attachmentC = zipC.read { storage.privilegedImportAttachment(it, "app", "C.zip") }
-            val attachmentD = zipD.read { storage.privilegedImportAttachment(it, "untrusted", null) }
-
-            assertThat(storage.resolveAttachmentTrustRoots()).containsOnly(
-                AttachmentTrustRoot(
-                    attachmentId = attachmentA,
-                    fileName = "A.jar",
-                    uploader = "app",
-                    trustRootId = attachmentA,
-                    trustRootFileName = "A.jar"
-                ),
-                AttachmentTrustRoot(
-                    attachmentId = attachmentB,
-                    fileName = "B.jar",
-                    uploader = "untrusted",
-                    trustRootId = attachmentA,
-                    trustRootFileName = "A.jar"
-                ),
-                AttachmentTrustRoot(
-                    attachmentId = attachmentC,
-                    fileName = "C.zip",
-                    uploader = "app",
-                    trustRootId = attachmentC,
-                    trustRootFileName = "C.zip"
-                ),
-                AttachmentTrustRoot(
-                    attachmentId = attachmentD,
-                    fileName = null,
-                    uploader = "untrusted",
-                    trustRootId = null,
-                    trustRootFileName = null
-                )
-            )
-        }
-    }
-
-    @Test
-    fun `resolveAttachmentTrustRoots attachments signed by blacklisted keys output without trust root fields filled in`() {
-        SelfCleaningDir().use { file ->
-
-            val aliasA = "batman"
-            val aliasB = "the joker"
-            val password = "nanananana batman"
-            file.path.generateKey(aliasA, password)
-            val keyB = file.path.generateKey(aliasB, password)
-
-            storage = NodeAttachmentService(
-                MetricRegistry(),
-                TestingNamedCacheFactory(),
-                database,
-                keysToBlacklist = listOf(keyB.hash.toString())
-            ).also {
-                database.transaction {
-                    it.start()
-                }
-            }
-            storage.servicesForResolution = services
-
-            val jarSignedByA = makeTestContractJar(file.path, "foo.bar.AnotherContract")
-            file.path.signJar(jarSignedByA.toAbsolutePath().toString(), aliasA, password)
-
-            val jarSignedByAB = makeTestContractJar(file.path, "foo.bar.DummyContract")
-            file.path.signJar(jarSignedByAB.toAbsolutePath().toString(), aliasA, password)
-            file.path.signJar(jarSignedByAB.toAbsolutePath().toString(), aliasB, password)
-
-            val attachmentA = jarSignedByA.read { storage.privilegedImportAttachment(it, "app", "A.jar") }
-            val attachmentB = jarSignedByAB.read { storage.privilegedImportAttachment(it, "untrusted", "B.jar") }
-
-            assertThat(storage.resolveAttachmentTrustRoots()).containsOnly(
-                AttachmentTrustRoot(
-                    attachmentId = attachmentA,
-                    fileName = "A.jar",
-                    uploader = "app",
-                    trustRootId = attachmentA,
-                    trustRootFileName = "A.jar"
-                ),
-                AttachmentTrustRoot(
-                    attachmentId = attachmentB,
-                    fileName = "B.jar",
-                    uploader = "untrusted",
-                    trustRootId = null,
-                    trustRootFileName = null
-                )
             )
         }
     }
